move base hcio callback function to hcio.py
